// src/chat-simulation/enums/chat-simuation.enum.ts
export enum AICharacterType {
  ACTOR = 'Actor',
  PANELIST = 'Panelist',
}

export enum EpisodeStatus {
  PENDING = 'Pending',
  POSTED = 'Posted',
  ONGOING = 'Ongoing',
  QUEUED = 'Queued',
  COMPLETED = 'Completed',
}

export enum ScheduleType {
  DAILY = 'Daily',
  WEEKLY = 'Weekly',
  MONTHLY = 'Monthly',
  YEARLY = 'Yearly',
  NEVER = 'Never',
}