// src/utilities/enum/env.enum.ts
export enum EnvType {
  PRODUCTION = 'production',
  DEVELOP = 'develop',
}

export enum eventEmitterType {
  CREATE_QUIZ = 'CREATE_QUIZ',
  LIVE_EPISODE = 'LIVE_EPISODE',
  GET_USER_UPDATE = 'GET_USER_UPDATE',
  ADD_QUESTIONS_TO_QUIZ = 'ADD_QUESTIONS_TO_QUIZ',
  CHANNELS_EPISODE = 'CHANNELS_EPISODE',
  CHANNEL_EVENTS = 'CHANNEL_EVENTS',
  EPISODE_EVENTS = 'EPISODE_EVENTS',
  CHANNEL_HOME_FEED = 'CHANNEL_HOME_FEED',
  CHANNEL = 'CHANNEL',
  SUBSCRIPTION_UPDATE = 'SUBSCRIPTION_UPDATE',
  PODCAST_GENERATED = 'PODCAST_GENERATED',
}